:truck: rename: 디렉토리 이름 customException -> exception 으로 변경
