+++ src/main/java/gift/exception/CustomException.java
@@ -1,4 +1,4 @@
-package gift.customException;
+package gift.exception;
 
 public interface CustomException {
