+++ src/main/java/gift/exception/InvalidIdException.java
@@ -1,4 +1,4 @@
-package gift.customException;
+package gift.exception;
 
 public class InvalidIdException extends RuntimeException implements CustomException {
     public InvalidIdException() {
