+++ src/main/java/gift/exception/NoSuchProductIdException.java
@@ -1,4 +1,4 @@
-package gift.customException;
+package gift.exception;
 
 public class NoSuchProductIdException extends RuntimeException implements CustomException{
     public NoSuchProductIdException(){
