+++ src/main/java/gift/exception/NullContentException.java
@@ -1,4 +1,4 @@
-package gift.customException;
+package gift.exception;
 
 public class NullContentException extends RuntimeException implements CustomException {
